# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/page": {
      "get": {
        "operationId": "UserController_pageUser",
        "parameters": [
          {
            "name": "nickName",
            "required": false,
            "in": "query",
            "description": "昵称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "description": "性别代码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genderName",
            "required": false,
            "in": "query",
            "description": "性别名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "birth",
            "required": false,
            "in": "query",
            "description": "生日",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mobile",
            "required": false,
            "in": "query",
            "description": "手机号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "description": "地址",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressDetail",
            "required": false,
            "in": "query",
            "description": "地址详情",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maritalStatus",
            "required": false,
            "in": "query",
            "description": "婚姻状况代码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maritalStatusName",
            "required": false,
            "in": "query",
            "description": "婚姻状况名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "邮箱",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "创建时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "修改时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "删除时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pinyin",
            "required": false,
            "in": "query",
            "description": "全拼",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "py",
            "required": false,
            "in": "query",
            "description": "拼音首字母",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userName",
            "required": false,
            "in": "query",
            "description": "用户名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "birthStart",
            "required": false,
            "in": "query",
            "description": "生日日期开始",
            "schema": {
              "example": "2000-10-10",
              "type": "string"
            }
          },
          {
            "name": "birthEnd",
            "required": false,
            "in": "query",
            "description": "生日日期结束",
            "schema": {
              "example": "2000-10-10",
              "type": "string"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建日期开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建日期结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "排序对象",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "example": {
                "userName": "ASC",
                "createdAt": "DESC"
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/UserPageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询用户列表",
        "tags": ["用户模块"]
      }
    },
    "/user/schema-generator-holder-page-user": {
      "post": {
        "deprecated": true,
        "operationId": "UserController__pageUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["用户模块"]
      }
    },
    "/user/option": {
      "get": {
        "operationId": "UserController_pageOptionUser",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "返回的列表字段",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/UserPageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询用户列表(可选字段)",
        "tags": ["用户模块"]
      }
    },
    "/user/schema-generator-holder-page-option-user": {
      "post": {
        "deprecated": true,
        "operationId": "UserController__pageOptionUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPageOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["用户模块"]
      }
    },
    "/user/find/{id}": {
      "get": {
        "operationId": "UserController_findUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按id查询用户",
        "tags": ["用户模块"]
      }
    },
    "/user/findRolesByUserId/{id}": {
      "get": {
        "operationId": "UserController_findRolesByUserId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FindRolesByUserIdResp"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按id查询角色",
        "tags": ["用户模块"]
      }
    },
    "/user/assignRolesToUser/{id}": {
      "post": {
        "operationId": "UserController_assignRolesToUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesToUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按id查询角色",
        "tags": ["用户模块"]
      }
    },
    "/user/batch/query": {
      "get": {
        "operationId": "UserController_findUserListByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "用“,”拼接的id",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserListRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按ids查询用户",
        "tags": ["用户模块"]
      }
    },
    "/user/create": {
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "创建用户",
        "tags": ["用户模块"]
      }
    },
    "/user/update/{id}": {
      "post": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "修改用户",
        "tags": ["用户模块"]
      }
    },
    "/user/delete/{id}": {
      "post": {
        "operationId": "UserController_removeUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "删除用户",
        "tags": ["用户模块"]
      }
    },
    "/user/batch/delete": {
      "post": {
        "operationId": "UserController_batchRemoveUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量删除用户",
        "tags": ["用户模块"]
      }
    },
    "/user/updateStatus/{id}": {
      "post": {
        "operationId": "UserController_updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "修改用户状态",
        "tags": ["用户模块"]
      }
    },
    "/user/batch/status": {
      "post": {
        "operationId": "UserController_batchUpdateUserStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量修改用户状态",
        "tags": ["用户模块"]
      }
    },
...
  },
  "info": {
    "title": "BACKED API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "head": {
            "$ref": "#/components/schemas/ResponseHeadDto"
          }
        },
        "required": ["head"]
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {}
      },
...
      "UserPageOptionDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数"
          },
          "keyword": {
            "type": "string",
            "description": "关键字",
            "example": ""
          },
          "fields": {
            "type": "string",
            "description": "返回的列表字段",
            "example": ""
          }
        }
      },
...
      "FindRolesByUserIdResp": {
        "type": "object",
        "properties": {
          "list": {
            "description": "角色id列表",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": ["list"]
      },
      "AssignRolesToUserDto": {
        "type": "object",
        "properties": {
          "roleIds": {
            "description": "角色id列表",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": ["roleIds"]
      },
      "UserListRespDto": {
        "type": "object",
        "properties": {
          "notFoundIds": {
            "description": "未找到的id数组",
            "example": "",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "list": {
            "description": "用户列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRespDto"
            }
          }
        },
        "required": ["list"]
      },
...
    }
  }
}